jobs/generate_headline_assessment/requirements.txt
# requirements.txt for generate-headline-assessment job

google-genai>=0.1.0
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage>=2.24.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
pandas==2.2.3
tenacity==9.0.0
db-dtypes>=1.1.0
